Typo fixes, moved ResponseContext to appropriate package and initial creation of the PropertiesManager
