+++ src/main/java/net/technolords/micro/command/CommandManager.java
@@ -7,7 +7,7 @@ import org.apache.camel.Exchange;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import net.technolords.micro.ResponseContext;
+import net.technolords.micro.processor.ResponseContext;
 
 public class CommandManager {
     private final Logger LOGGER = LoggerFactory.getLogger(getClass());
