+++ src/main/java/net/technolords/micro/command/StopCommand.java
@@ -6,7 +6,7 @@ import org.apache.camel.Exchange;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import net.technolords.micro.ResponseContext;
+import net.technolords.micro.processor.ResponseContext;
 
 public class StopCommand {
     private static final Logger LOGGER = LoggerFactory.getLogger(StopCommand.class);
